Move complexity tests to their own module.
